requirements.txt
asgiref==3.8.1
attrs==24.1.0
Django==5.0.8
django-debug-toolbar==3.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.22.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
pep8-naming==0.13.2
pillow==10.4.0
psycopg2-binary==2.9.9
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.19.1
setuptools==72.1.0
sqlparse==0.5.1
uritemplate==4.1.1
